backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
aiosmtplib==3.0.2
email-validator==2.2.0
scikit-learn==1.5.2
numpy==2.1.3
